Refactor to move gpu_metrics struct and serialization to another file
