+++ projects/rocprofiler-systems/source/lib/core/trace_cache/rocpd_post_processing.cpp
@@ -24,6 +24,7 @@
 #include "agent_manager.hpp"
 #include "config.hpp"
 #include "debug.hpp"
+#include "gpu_metrics.hpp"
 #include "library/thread_info.hpp"
 #include "node_info.hpp"
 #include "rocpd/data_processor.hpp"
@@ -404,174 +405,8 @@ rocpd_post_processing::get_pmc_event_with_sample_callback() const
 postprocessing_callback
 rocpd_post_processing::get_amd_smi_sample_callback() const
 {
-    struct gpu_metrics_t
-    {
-        // VCN metrics
-        std::vector<uint16_t> vcn_activity;  // Device-level VCN (when supported)
-        std::vector<std::vector<uint16_t>>
-            vcn_busy;  // XCP-level VCN (per-XCP organization)
-
-        // JPEG metrics
-        std::vector<uint16_t> jpeg_activity;  // Device-level JPEG (when supported)
-        std::vector<std::vector<uint16_t>>
-            jpeg_busy;  // XCP-level JPEG (per-XCP organization)
-
-        // XGMI metrics
-        uint16_t              xgmi_link_width = 0;
-        uint16_t              xgmi_link_speed = 0;
-        std::vector<uint64_t> xgmi_read_data_acc;
-        std::vector<uint64_t> xgmi_write_data_acc;
-
-        // PCIe metrics
-        uint16_t pcie_link_width     = 0;
-        uint16_t pcie_link_speed     = 0;
-        uint64_t pcie_bandwidth_acc  = 0;
-        uint64_t pcie_bandwidth_inst = 0;
-    };
-
-    auto deserialize_gpu_metrics = [](const std::vector<uint8_t>& serialized_data,
-                                      gpu_metrics_t& result, bool is_vcn_enabled,
-                                      bool is_jpeg_enabled, bool is_xgmi_enabled,
-                                      bool is_pcie_enabled, bool& _is_vcn_supported,
-                                      bool& _is_jpeg_supported) {
-        // - Flags (1 byte): bit 0: is_vcn_supported, bit 1: is_jpeg_supported
-        // - counts (4 bytes): vcn_count, jpeg_count, xgmi_read_count, xgmi_write_count
-        // - conditional data based on settings (passed as parameters)
-        if(serialized_data.empty())
-        {
-            throw std::runtime_error("Invalid serialized data: insufficient header size");
-            // return;  // Empty data, nothing to deserialize
-        }
-        size_t offset = 0;
-
-        auto deserialize_uint8 = [](const std::vector<uint8_t>& data, size_t& _offset) {
-            if(_offset >= data.size())
-                throw std::runtime_error("Invalid serialized data: unexpected end");
-            return data[_offset++];
-        };
-
-        auto deserialize_uint16 = [](const std::vector<uint8_t>& data, size_t& _offset) {
-            if(_offset + 1 >= data.size())
-                throw std::runtime_error("Invalid serialized data: unexpected end");
-            uint16_t value = static_cast<uint16_t>(data[_offset]) |
-                             (static_cast<uint16_t>(data[_offset + 1]) << 8);
-            _offset += 2;
-            return value;
-        };
-
-        auto deserialize_uint64 = [](const std::vector<uint8_t>& data, size_t& _offset) {
-            if(_offset + 7 >= data.size())
-                throw std::runtime_error("Invalid serialized data: unexpected end");
-            uint64_t value = 0;
-            for(int i = 0; i < 8; ++i)
-                value |= (static_cast<uint64_t>(data[_offset + i]) << (i * 8));
-            _offset += 8;
-            return value;
-        };
-
-        auto deserialize_uint16_vector = [&](const std::vector<uint8_t>& data,
-                                             size_t& _offset, uint8_t count) {
-            std::vector<uint16_t> values;
-            values.reserve(count);
-            for(uint8_t i = 0; i < count; ++i)
-                values.push_back(deserialize_uint16(data, _offset));
-            return values;
-        };
-
-        auto deserialize_uint64_vector = [&](const std::vector<uint8_t>& data,
-                                             size_t& _offset, uint8_t count) {
-            std::vector<uint64_t> values;
-            values.reserve(count);
-            for(uint8_t i = 0; i < count; ++i)
-                values.push_back(deserialize_uint64(data, _offset));
-            return values;
-        };
-
-        // Deserialize capability flags
-        uint8_t _flags     = deserialize_uint8(serialized_data, offset);
-        _is_vcn_supported  = (_flags & 0x01) != 0;
-        _is_jpeg_supported = (_flags & 0x02) != 0;
-
-        // Deserialize counts
-        uint8_t vcn_count        = deserialize_uint8(serialized_data, offset);
-        uint8_t jpeg_count       = deserialize_uint8(serialized_data, offset);
-        uint8_t vcn_xcp_count    = deserialize_uint8(serialized_data, offset);
-        uint8_t jpeg_xcp_count   = deserialize_uint8(serialized_data, offset);
-        uint8_t xgmi_read_count  = deserialize_uint8(serialized_data, offset);
-        uint8_t xgmi_write_count = deserialize_uint8(serialized_data, offset);
-
-        // Deserialize per-XCP sizes
-        std::vector<uint8_t> vcn_xcp_sizes;
-        std::vector<uint8_t> jpeg_xcp_sizes;
-        for(uint8_t i = 0; i < vcn_xcp_count; ++i)
-            vcn_xcp_sizes.push_back(deserialize_uint8(serialized_data, offset));
-        for(uint8_t i = 0; i < jpeg_xcp_count; ++i)
-            jpeg_xcp_sizes.push_back(deserialize_uint8(serialized_data, offset));
-
-        // Deserialize VCN data and reconstruct structure
-        if(is_vcn_enabled && vcn_count > 0)
-        {
-            auto flat_data =
-                deserialize_uint16_vector(serialized_data, offset, vcn_count);
-            if(_is_vcn_supported)
-            {
-                result.vcn_activity = flat_data;
-            }
-            else
-            {
-                // Per-XCP: split flat data according to XCP sizes into vcn_busy
-                size_t flat_offset = 0;
-                for(uint8_t xcp_size : vcn_xcp_sizes)
-                {
-                    std::vector<uint16_t> xcp_data(flat_data.begin() + flat_offset,
-                                                   flat_data.begin() + flat_offset +
-                                                       xcp_size);
-                    result.vcn_busy.push_back(xcp_data);
-                    flat_offset += xcp_size;
-                }
-            }
-        }
-
-        // Deserialize JPEG data and reconstruct structure
-        if(is_jpeg_enabled && jpeg_count > 0)
-        {
-            auto flat_data =
-                deserialize_uint16_vector(serialized_data, offset, jpeg_count);
-            if(_is_jpeg_supported)
-            {
-                result.jpeg_activity = flat_data;
-            }
-            else
-            {
-                // Per-XCP: split flat data according to XCP sizes into jpeg_busy
-                size_t flat_offset = 0;
-                for(uint8_t xcp_size : jpeg_xcp_sizes)
-                {
-                    std::vector<uint16_t> xcp_data(flat_data.begin() + flat_offset,
-                                                   flat_data.begin() + flat_offset +
-                                                       xcp_size);
-                    result.jpeg_busy.push_back(xcp_data);
-                    flat_offset += xcp_size;
-                }
-            }
-        }
-        if(is_xgmi_enabled)
-        {
-            result.xgmi_link_width = deserialize_uint16(serialized_data, offset);
-            result.xgmi_link_speed = deserialize_uint16(serialized_data, offset);
-            result.xgmi_read_data_acc =
-                deserialize_uint64_vector(serialized_data, offset, xgmi_read_count);
-            result.xgmi_write_data_acc =
-                deserialize_uint64_vector(serialized_data, offset, xgmi_write_count);
-        }
-        if(is_pcie_enabled)
-        {
-            result.pcie_link_width     = deserialize_uint16(serialized_data, offset);
-            result.pcie_link_speed     = deserialize_uint16(serialized_data, offset);
-            result.pcie_bandwidth_acc  = deserialize_uint64(serialized_data, offset);
-            result.pcie_bandwidth_inst = deserialize_uint64(serialized_data, offset);
-        }
-    };
+    // Use the shared gpu_metrics_t from core/gpu_metrics.hpp
+    using gpu_metrics_t = gpu::gpu_metrics_t;
 
     return [&](const storage_parsed_type_base& parsed) {
         auto _amd_smi = static_cast<const struct amd_smi_sample&>(parsed);
@@ -644,9 +479,10 @@ rocpd_post_processing::get_amd_smi_sample_callback() const
         gpu_metrics_t gpu_metrics;
         bool          is_vcn_activity_supported  = false;
         bool          is_jpeg_activity_supported = false;
-        deserialize_gpu_metrics(_amd_smi.gpu_activity, gpu_metrics, is_vcn_enabled,
-                                is_jpeg_enabled, is_xgmi_enabled, is_pcie_enabled,
-                                is_vcn_activity_supported, is_jpeg_activity_supported);
+        gpu::deserialize_gpu_metrics(_amd_smi.gpu_activity, gpu_metrics, is_vcn_enabled,
+                                     is_jpeg_enabled, is_xgmi_enabled, is_pcie_enabled,
+                                     is_vcn_activity_supported,
+                                     is_jpeg_activity_supported);
 
         // Insert VCN and JPEG activity metrics
         auto insert_decode_vector_metrics = [&](auto category, bool _is_enabled,
