+++ projects/rocprofiler-systems/tests/rocpd-validation-rules/gpu-connect/amd-smi-rules.json
@@ -2,7 +2,7 @@
     "required_tables": [
         {
             "min_rows": 1,
-            "name_prefix": "rocpd_info_pmc_",
+            "name_prefix": "rocpd_info_pmc",
             "required_columns": [
                 "agent_id",
                 "target_arch",
@@ -15,15 +15,15 @@
             "validation_queries": [
                 {
                     "comparison": "greater_than",
-                    "description": "Check for jpeg_activity amd-smi metrics",
-                    "error_message": "Did not find jpeg_activity in amd-smi metrics",
+                    "description": "Check for Xgmi amd-smi metrics",
+                    "error_message": "Did not find Xgmi data in amd-smi metrics",
                     "expected_result": 1,
                     "query": "SELECT COUNT(*) as count FROM {table_name} WHERE symbol LIKE 'Xgmi%'"
                 },
                 {
                     "comparison": "greater_than",
-                    "description": "Check for jpeg_activity amd-smi metrics",
-                    "error_message": "Did not find jpeg_activity in amd-smi metrics",
+                    "description": "Check for Pcie amd-smi metrics",
+                    "error_message": "Did not find Pcie data in amd-smi metrics",
                     "expected_result": 1,
                     "query": "SELECT COUNT(*) as count FROM {table_name} WHERE symbol LIKE 'Pcie%'"
                 }
@@ -31,7 +31,7 @@
         },
         {
             "min_rows": 500,
-            "name_prefix": "rocpd_pmc_event_",
+            "name_prefix": "rocpd_pmc_event",
             "required_columns": [
                 "event_id",
                 "pmc_id",
@@ -41,23 +41,58 @@
                 {
                     "comparison": "greater_than",
                     "description": "Check for amd-smi monitoring busy times",
-                    "error_message": "Less than expected number of captured amd-smi mm-busy samples!",
+                    "error_message": "Less than expected number of captured amd-smi xgmi link speed samples!",
                     "expected_result": 100,
-                    "query": "SELECT COUNT(*) as count FROM {table_name} event JOIN rocpd_info_pmc info ON event.pmc_id = info.id WHERE info.name LIKE 'device_xgmi%'"
+                    "query": "SELECT COUNT(*) as count FROM {table_name} event JOIN rocpd_info_pmc info ON event.pmc_id = info.id WHERE info.name = 'device_xgmi_link_speed'"
                 },
                 {
                     "comparison": "greater_than",
-                    "description": "Check for amd-smi monitoring GPU memory usage",
-                    "error_message": "Less than expected number of captured amd-smi memory-usage samples!",
-                    "expected_result": 50,
-                    "query": "SELECT COUNT(*) as count FROM {table_name} event JOIN rocpd_info_pmc info ON event.pmc_id = info.id WHERE info.name = 'device_memory_usage'"
+                    "description": "Check for amd-smi monitoring busy times",
+                    "error_message": "Less than expected number of captured amd-smi xgmi link width samples!",
+                    "expected_result": 100,
+                    "query": "SELECT COUNT(*) as count FROM {table_name} event JOIN rocpd_info_pmc info ON event.pmc_id = info.id WHERE info.name = 'device_xgmi_link_width'"
                 },
                 {
                     "comparison": "greater_than",
-                    "description": "Check for amd-smi monitoring JPEG activity",
-                    "error_message": "Less than expected activity in amd-smi jpeg-activity samples!",
-                    "expected_result": 50,
-                    "query": "SELECT COUNT(*) as count FROM {table_name} event JOIN rocpd_info_pmc info ON event.pmc_id = info.id WHERE info.name LIKE 'device_jpeg_activity_%' and event.value > 0"
+                    "description": "Check for amd-smi monitoring busy times",
+                    "error_message": "Less than expected number of captured amd-smi xgmi read data samples!",
+                    "expected_result": 100,
+                    "query": "SELECT COUNT(*) as count FROM {table_name} event JOIN rocpd_info_pmc info ON event.pmc_id = info.id WHERE info.name LIKE 'device_xgmi_read_data%'"
+                },
+                {
+                    "comparison": "greater_than",
+                    "description": "Check for amd-smi monitoring busy times",
+                    "error_message": "Less than expected number of captured amd-smi xgmi write data samples!",
+                    "expected_result": 100,
+                    "query": "SELECT COUNT(*) as count FROM {table_name} event JOIN rocpd_info_pmc info ON event.pmc_id = info.id WHERE info.name LIKE 'device_xgmi_write_data%'"
+                },
+                {
+                    "comparison": "greater_than",
+                    "description": "Check for amd-smi monitoring busy times",
+                    "error_message": "Less than expected number of captured amd-smi pcie bandwidth instantaneous samples!",
+                    "expected_result": 100,
+                    "query": "SELECT COUNT(*) as count FROM {table_name} event JOIN rocpd_info_pmc info ON event.pmc_id = info.id WHERE info.name = 'device_pcie_bandwidth_inst'"
+                },
+                {
+                    "comparison": "greater_than",
+                    "description": "Check for amd-smi monitoring busy times",
+                    "error_message": "Less than expected number of captured amd-smi pcie bandwidth accumulated samples!",
+                    "expected_result": 100,
+                    "query": "SELECT COUNT(*) as count FROM {table_name} event JOIN rocpd_info_pmc info ON event.pmc_id = info.id WHERE info.name = 'device_pcie_bandwidth_acc'"
+                },
+                {
+                    "comparison": "greater_than",
+                    "description": "Check for amd-smi monitoring busy times",
+                    "error_message": "Less than expected number of captured amd-smi pcie link speed samples!",
+                    "expected_result": 100,
+                    "query": "SELECT COUNT(*) as count FROM {table_name} event JOIN rocpd_info_pmc info ON event.pmc_id = info.id WHERE info.name = 'device_pcie_link_speed'"
+                },
+                {
+                    "comparison": "greater_than",
+                    "description": "Check for amd-smi monitoring busy times",
+                    "error_message": "Less than expected number of captured amd-smi pcie link width samples!",
+                    "expected_result": 100,
+                    "query": "SELECT COUNT(*) as count FROM {table_name} event JOIN rocpd_info_pmc info ON event.pmc_id = info.id WHERE info.name = 'device_pcie_link_width'"
                 }
             ]
         }
